% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_publish_to_ah}
\alias{addin_publish_to_ah}
\title{Addin for publish_to_ah for publishing Distill Articles to the Knowledge Repo}
\usage{
addin_publish_to_ah()
}
\description{
Addin for publish_to_ah for publishing Distill Articles to the Knowledge Repo
}
\keyword{addins}
