% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_summary.R
\name{brand_hooks}
\alias{brand_hooks}
\title{Second Generation Change for Block Knitting}
\usage{
brand_hooks(x, options)
}
\arguments{
\item{x}{the input}

\item{options}{more information on options}
}
\description{
Second Generation Change for Block Knitting
}
