% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_report.R
\name{brand_report}
\alias{brand_report}
\title{Generating an EA Report Automatically}
\usage{
brand_report(
  toc = TRUE,
  toc_float = TRUE,
  df_print = "paged",
  highlight = "tango",
  code_folding = "hide",
  fig_width = 10,
  fig_height = 6,
  fig_retina = 1,
  code_download = TRUE,
  analyst,
  email,
  draft = FALSE,
  audience = "internal",
  ...
)
}
\arguments{
\item{toc}{default is \code{TRUE}}

\item{toc_float}{default is \code{TRUE}}

\item{df_print}{default is \code{"paged"}}

\item{highlight}{default is \code{"tango"}}

\item{code_folding}{default is \code{"hide"}}

\item{fig_width}{default is \code{10}}

\item{fig_height}{default is \code{6}}

\item{fig_retina}{default is \code{1}}

\item{code_download}{default is \code{TRUE}}

\item{analyst}{the analyst name to be used}

\item{email}{the email to be included in the footer}

\item{draft}{if "DRAFT" is to be printed}

\item{audience}{the targeted audience for the output one of \code{"internal"}
or \code{"client"}}

\item{...}{any additional parameters to pass to \code{html_document2}}
}
\description{
This function will automatically compile an EA
    style report without having to tweak the internals of the yaml header
}
