% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_sql}
\alias{check_sql}
\title{Helper function to check that at minimum the sql
template contains \code{SELECT} and \code{FROM}
to verify that a valid query is being submitted}
\usage{
check_sql(template)
}
\arguments{
\item{template}{the template to check}
}
\description{
Helper function to check that at minimum the sql
template contains \code{SELECT} and \code{FROM}
to verify that a valid query is being submitted
}
