% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_dropdown.R
\name{knit_dropdown}
\alias{knit_dropdown}
\title{Add Dropdown Box to Knitr Options}
\usage{
knit_dropdown(before, options, envir)
}
\arguments{
\item{before}{before option}

\item{options}{the options option}

\item{envir}{the environment}
}
\description{
adds an option to rmarkdown code chunks to have a dropdown.
    This will allow users to more easily hide entire code chunks if desired.
}
\examples{
\dontrun{# In an Rmarkdown Code Chunk Put the Following:

knitr::knit_hooks$set(dropdown = knit_dropdown)

}


}
