% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_summary.R
\name{knit_summary}
\alias{knit_summary}
\title{Knit Summary Block}
\usage{
knit_summary(before, options, envir)
}
\arguments{
\item{before}{before the code chunk is executed}

\item{options}{any additional options}

\item{envir}{the environment}
}
\description{
Knit Summary Block
}
