% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project_template.R
\name{make_project_templates}
\alias{make_project_templates}
\title{make project template}
\usage{
make_project_templates(
  full = T,
  shiny = FALSE,
  location = here::here(),
  add_report = T
)
}
\arguments{
\item{full}{default is T, build the reduced structure or full structure}

\item{shiny}{with default of \code{FALSE}}

\item{location}{the path you wish to put these documents}

\item{add_report}{default is \code{TRUE}}
}
\description{
build a project template that has the associated directories
    commonly used
}
\concept{project template}
