% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repo.R
\name{prep_repo}
\alias{prep_repo}
\title{Prep A Project Repository}
\usage{
prep_repo(location = here::here())
}
\arguments{
\item{location}{a filepath, the location to write the file
with a default of \code{here::here()}}
}
\description{
This function applies good practices and writes several
invisible files to a repository in order to ensure that
secrets are not shared and that large data files are not
stored.
}
