% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repo.R
\name{write_attributes}
\alias{write_attributes}
\title{Write Attributes}
\usage{
write_attributes(location = here::here())
}
\arguments{
\item{location}{a filepath, the location to write the file
with a default of \code{here::here()}}
}
\description{
Adds a .gitattributes file to a directory
}
