% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repo.R
\name{write_dockerignore}
\alias{write_dockerignore}
\title{Write Dockerignore}
\usage{
write_dockerignore(location = here::here())
}
\arguments{
\item{location}{a filepath, the location to write the file
with a default of \code{here::here()}}
}
\description{
Adds a .dockerignore file to a directory
}
