% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ea_report_footer.R
\name{write_ea_report_footer}
\alias{write_ea_report_footer}
\title{Footer Generation for EA Reports}
\usage{
write_ea_report_footer(
  location = NULL,
  author_email = rmarkdown::metadata$email,
  author_name = NULL
)
}
\arguments{
\item{location}{default value of \code{NULL} for where to generate the document}

\item{author_email}{the email to use for the author. If not specified,
Then the rmarkdown email will be used if available. If not email specified
then the footer will omit this entire line.}
}
\description{
Footer Generation for EA Reports
}
\keyword{internal}
