% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_repo.R
\name{write_gitignore}
\alias{write_gitignore}
\title{Write Gitignore}
\usage{
write_gitignore(location = here::here())
}
\arguments{
\item{location}{a filepath, the location to write the file
with a default of \code{here::here()}}
}
\description{
Adds a .gitignore file to a directory
}
